requirements.txt
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
crewai>=0.28.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
opencv-python>=4.7.0
scikit-learn>=1.2.0
duckduckgo-search>=3.8.0
reportlab>=4.0.0
python-dotenv>=1.0.0
exif>=1.6.0
numpy>=1.21.0
matplotlib>=3.6.0
requests>=2.28.0
